// my-client/src/app/interfaces/cart.ts
export interface CartItem {
    productId: string;
    category_id: string;
    Name: string;
    Price: number;
    Image: string;
    Description: string;
    Origin: string;
    Uses: string;
    Store: string;
    Quantity: number;     
    cartQuantity: number; 
    Create_date: Date | string;
    Dimension: string;
    Story: string;
    ProductCare: string;
    ShippingReturn: string;
}